fix(recovery): Fix recovery with change of size of `AdrenalineConfig`

+++ cef/newrecovery/menu.h
@@ -25,7 +25,7 @@ typedef struct {
 	void (* function)(int sel);
 	char **options;
 	int size_options;
-	int *value;
+	u8 *value;
 } Entry;
 
 typedef struct {
